refactor entity classes out to separate files; update all references
