+++ ContentManagement/Unifi.cs
@@ -7,6 +7,8 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Windows;
+using ContentManagement.Entities;
+using Parameter = ContentManagement.Entities.Parameter;
 
 namespace ContentManagement {
     class Unifi {
@@ -121,7 +123,7 @@ namespace ContentManagement {
             List<string> familyTypes = new List<string>();
 
             // Loop through all parameters
-            foreach (Unifi.Parameter p in content.Parameters) {
+            foreach (Parameter p in content.Parameters) {
                 // Pass parameter values to Content object
                 if (p.TypeName != "") { familyTypes.Add(p.TypeName); }
             }
@@ -211,359 +213,6 @@ namespace ContentManagement {
                     return writer.ToString();
                 }
             }
-        }
-
-        public partial class Content {
-            // Pass values from Manufacturer and Model parameters to object
-            public string Manufacturer { get; set; }
-            public string Model { get; set; }
-
-            [JsonProperty("ActiveRevisionId")]
-            public Guid ActiveRevisionId { get; set; }
-
-            [JsonProperty("AggregateRating")]
-            public object AggregateRating { get; set; }
-
-            [JsonProperty("ApprovalStatus")]
-            public long ApprovalStatus { get; set; }
-
-            [JsonProperty("ApprovedRevisionId")]
-            public Guid ApprovedRevisionId { get; set; }
-
-            [JsonProperty("Brands")]
-            public object[] Brands { get; set; }
-
-            [JsonProperty("Category")]
-            public string Category { get; set; }
-
-            [JsonProperty("Channels")]
-            public object[] Channels { get; set; }
-
-            [JsonProperty("CreatedDate")]
-            public string CreatedDate { get; set; }
-
-            [JsonProperty("CreatorUsername")]
-            public string CreatorUsername { get; set; }
-
-            [JsonProperty("CurrentRevisionId")]
-            public Guid CurrentRevisionId { get; set; }
-
-            [JsonProperty("Downloads")]
-            public long Downloads { get; set; }
-
-            public List<string> FamilyTypes { get; set; }
-
-            [JsonProperty("Favorites")]
-            public object[] Favorites { get; set; }
-
-            [JsonProperty("FileImageBytes")]
-            public object FileImageBytes { get; set; }
-
-            [JsonProperty("FileType")]
-            public long FileType { get; set; }
-
-            [JsonProperty("Filename")]
-            public string Filename { get; set; }
-
-            [JsonProperty("FillPatternType")]
-            public object FillPatternType { get; set; }
-
-            [JsonProperty("HasCustomPreviewImage")]
-            public bool HasCustomPreviewImage { get; set; }
-
-            [JsonProperty("HasTypeCatalog")]
-            public bool HasTypeCatalog { get; set; }
-
-            [JsonProperty("LastModifiedDate")]
-            public string LastModifiedDate { get; set; }
-
-            [JsonProperty("Libraries")]
-            public Library[] Libraries { get; set; }
-
-            [JsonProperty("LocalPath")]
-            public object LocalPath { get; set; }
-
-            [JsonProperty("MaterialClass")]
-            public object MaterialClass { get; set; }
-
-            [JsonProperty("MeasurementSystem")]
-            public long MeasurementSystem { get; set; }
-
-            [JsonProperty("Measurements")]
-            public object[] Measurements { get; set; }
-
-            [JsonProperty("NextRevisionNumber")]
-            public object NextRevisionNumber { get; set; }
-
-            [JsonProperty("OriginalDateAdded")]
-            public string OriginalDateAdded { get; set; }
-
-            [JsonProperty("ParseStatus")]
-            public long ParseStatus { get; set; }
-
-            [JsonProperty("PreviewImageUrl")]
-            public string PreviewImageUrl { get; set; }
-
-            [JsonProperty("Ratings")]
-            public object[] Ratings { get; set; }
-
-            [JsonProperty("RepositoryFileId")]
-            public Guid RepositoryFileId { get; set; }
-
-            [JsonProperty("RepositoryNumber")]
-            public object RepositoryNumber { get; set; }
-
-            [JsonProperty("RevisionNumber")]
-            public object RevisionNumber { get; set; }
-
-            [JsonProperty("Revisions")]
-            public Revision[] Revisions { get; set; }
-
-            [JsonProperty("RevitYear")]
-            public long RevitYear { get; set; }
-
-            [JsonProperty("Size")]
-            public long Size { get; set; }
-
-            [JsonProperty("Tags")]
-            public Tag[] Tags { get; set; }
-
-            [JsonProperty("Title")]
-            public string Title { get; set; }
-
-            [JsonProperty("TypeCatalogBytes")]
-            public object TypeCatalogBytes { get; set; }
-
-            [JsonProperty("UserRating")]
-            public long UserRating { get; set; }
-
-            [JsonProperty("Parameters")]
-            public Parameter[] Parameters { get; set; }
-        }
-
-        public partial class Library {
-            [JsonProperty("AccessibleUserGroups")]
-            public object[] AccessibleUserGroups { get; set; }
-
-            [JsonProperty("AccessibleUsers")]
-            public object[] AccessibleUsers { get; set; }
-
-            [JsonProperty("AdminUserGroups")]
-            public object[] AdminUserGroups { get; set; }
-
-            [JsonProperty("AdminUsers")]
-            public object[] AdminUsers { get; set; }
-
-            [JsonProperty("CompanyId")]
-            public Guid CompanyId { get; set; }
-
-            [JsonProperty("CompanyName")]
-            public string CompanyName { get; set; }
-
-            [JsonProperty("DateCreated")]
-            public string DateCreated { get; set; }
-
-            [JsonProperty("IsProtected")]
-            public bool IsProtected { get; set; }
-
-            [JsonProperty("id")]
-            public Guid Id { get; set; }
-
-            [JsonProperty("LibraryId")]
-            private Guid LibraryId {
-                set { Id = value; }
-            }
-
-            [JsonProperty("LibraryType")]
-            public long LibraryType { get; set; }
-
-            [JsonProperty("name")]
-            public string Name { get; set; }
-
-            [JsonProperty("Name")]
-            private string Name2 {
-                set { Name = value; }
-            }
-
-            [JsonProperty("RepositoryId")]
-            public Guid RepositoryId { get; set; }
-        }
-
-        public partial class Parameter {
-            [JsonProperty("TypeName")]
-            public string TypeName { get; set; }
-
-            [JsonProperty("ParameterType")]
-            public long ParameterType { get; set; }
-
-            [JsonProperty("StorageType")]
-            public long StorageType { get; set; }
-
-            [JsonProperty("DisplayUnitType")]
-            public long? DisplayUnitType { get; set; }
-
-            [JsonProperty("IsDeterminedByFormula")]
-            public bool IsDeterminedByFormula { get; set; }
-
-            [JsonProperty("GUID")]
-            public string Guid { get; set; }
-
-            [JsonProperty("CanAssignFormula")]
-            public bool CanAssignFormula { get; set; }
-
-            [JsonProperty("IsBaseVersion")]
-            public bool IsBaseVersion { get; set; }
-
-            [JsonProperty("FamilyCategory")]
-            public string FamilyCategory { get; set; }
-
-            [JsonProperty("FileVersionId")]
-            public Guid FileVersionId { get; set; }
-
-            [JsonProperty("Name")]
-            public string Name { get; set; }
-
-            [JsonProperty("BuiltInParameter")]
-            public long BuiltInParameter { get; set; }
-
-            [JsonProperty("RevitYear")]
-            public long RevitYear { get; set; }
-
-            [JsonProperty("IsReporting")]
-            public bool IsReporting { get; set; }
-
-            [JsonProperty("FileRevisionId")]
-            public Guid FileRevisionId { get; set; }
-
-            [JsonProperty("ParameterGroup")]
-            public long ParameterGroup { get; set; }
-
-            [JsonProperty("Visible")]
-            public bool Visible { get; set; }
-
-            [JsonProperty("Value")]
-            public string Value { get; set; }
-
-            [JsonProperty("FileId")]
-            public Guid FileId { get; set; }
-
-            [JsonProperty("NumericValue")]
-            public string NumericValue { get; set; }
-
-            [JsonProperty("SetByTypeCatalog")]
-            public bool SetByTypeCatalog { get; set; }
-
-            [JsonProperty("IsReadOnly")]
-            public bool IsReadOnly { get; set; }
-
-            [JsonProperty("ElementId")]
-            public long ElementId { get; set; }
-
-            [JsonProperty("IsInstance")]
-            public bool IsInstance { get; set; }
-        }
-
-        public partial class Revision {
-            [JsonProperty("BaseFileVersions")]
-            public BaseFileVersion[] BaseFileVersions { get; set; }
-
-            [JsonProperty("Created")]
-            public string Created { get; set; }
-
-            [JsonProperty("FileRevisionId")]
-            public Guid FileRevisionId { get; set; }
-
-            [JsonProperty("FillPatternType")]
-            public object FillPatternType { get; set; }
-
-            [JsonProperty("MaterialClass")]
-            public object MaterialClass { get; set; }
-
-            [JsonProperty("Notes")]
-            public string Notes { get; set; }
-
-            [JsonProperty("RevisionNumber")]
-            public long RevisionNumber { get; set; }
-
-            [JsonProperty("Status")]
-            public long Status { get; set; }
-
-            [JsonProperty("Username")]
-            public string Username { get; set; }
-        }
-
-        public partial class BaseFileVersion {
-            [JsonProperty("FileVersionId")]
-            public Guid FileVersionId { get; set; }
-
-            [JsonProperty("FillPatternType")]
-            public object FillPatternType { get; set; }
-
-            [JsonProperty("IsBaseVersion")]
-            public bool IsBaseVersion { get; set; }
-
-            [JsonProperty("MaterialClass")]
-            public object MaterialClass { get; set; }
-
-            [JsonProperty("OriginalFilename")]
-            public string OriginalFilename { get; set; }
-
-            [JsonProperty("RepositoryNumber")]
-            public long RepositoryNumber { get; set; }
-
-            [JsonProperty("RevitYear")]
-            public long RevitYear { get; set; }
-        }
-
-        public partial class Tag {
-            [JsonProperty("RepositoryNumber")]
-            public long RepositoryNumber { get; set; }
-
-            [JsonProperty("TagId")]
-            public Guid TagId { get; set; }
-
-            [JsonProperty("TagString")]
-            public string TagString { get; set; }
-        }
-
-        public partial class Batch {
-            //[JsonProperty("Summary")]
-            //public Summary Summary { get; set; }
-
-            [JsonProperty("Details")]
-            public object[] Details { get; set; }
-
-            //[JsonProperty("FileStatus")]
-            //public FileStatus[] FileStatus { get; set; }
-
-            [JsonProperty("BatchId")]
-            public Guid BatchId { get; set; }
-        }
-
-        public partial class BatchStatus {
-            [JsonProperty("resultDetails")]
-            public string ResultDetails { get; set; }
-
-            [JsonProperty("TotalFiles")]
-            public long TotalFiles { get; set; }
-
-            [JsonProperty("OkFiles")]
-            public long OkFiles { get; set; }
-
-            [JsonProperty("PendingFiles")]
-            public long PendingFiles { get; set; }
-
-            [JsonProperty("FailedFiles")]
-            public long FailedFiles { get; set; }
-
-            [JsonProperty("TotalOperations")]
-            public long TotalOperations { get; set; }
-
-            [JsonProperty("CompletedOperations")]
-            public long CompletedOperations { get; set; }
-
-            [JsonProperty("Error")]
-            public string Error { get; set; }
         }
     }
 }
